+++ cache/src/main/kotlin/world/gregs/voidps/cache/definition/data/ObjectDefinitionFull.kt
@@ -80,7 +80,7 @@ data class ObjectDefinitionFull(
     var anInt3020: Int = 256,
     var aBoolean2992: Boolean = false,
     var anInt2975: Int = 0,
-    override var params: Map<Long, Any>? = null,
+    override var params: Map<Int, Any>? = null,
     override var stringId: String = "",
     override var extras: Map<String, Any>? = null
 ) : Definition, Transforms, Recolourable, ColourPalette, Parameterized, Extra {
